+++ lib/app_router.dart
@@ -1,0 +1,17 @@
+import 'package:ecomeal/views/startup/startup_view.dart';
+import 'package:flutter/material.dart';
+
+import 'constants/routes.dart';
+
+/// [AppRouter]
+/// This the base router to handle routes
+class AppRouter {
+  static Route<dynamic> generateRoute(RouteSettings settings) {
+    switch (settings.name) {
+      case RoutePaths.startup:
+        return MaterialPageRoute(builder: (_) => const StartupView());
+      default:
+        return MaterialPageRoute(builder: (_) => Container());
+    }
+  }
+}