+++ lib/constants/assets.dart
@@ -1,0 +1,7 @@
+class ImagePaths {
+  ImagePaths._();
+}
+
+class IconPaths {
+  IconPaths._();
+}