+++ lib/constants/enums.dart
@@ -1,0 +1,8 @@
+// All app's enums here
+
+/// [ViewModelState]
+/// Enum used to manage a viewmodel's state directly from the viewmodel.
+enum ViewModelState {
+  busy,
+  idle,
+}