+++ lib/viewmodels/startup/startup_viewmodel.dart
@@ -1,0 +1,4 @@
+import 'package:ecomeal/viewmodels/base.viewmodel.dart';
+
+class StartupViewModel extends BaseViewModel {
+}