+++ lib/views/startup/startup_view.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class StartupView extends StatelessWidget {
+  const StartupView({
+    super.key
+  });
+
+  @override
+  Widget build(BuildContext context){
+    return const Placeholder(color: Colors.white);
+  }
+}